+++ examples/macos/ViewController.h
@@ -5,4 +5,11 @@
 @interface ViewController : NSViewController <MTKViewDelegate>
 - (void)loadURL:(NSURL *)url;
 - (IBAction)toggleDebugMode:(id)sender;
+- (IBAction)togglePunctualLights:(id)sender;
+- (IBAction)toggleDefaultLight:(id)sender;
+- (IBAction)setBackgroundDark:(id)sender;
+- (IBAction)setBackgroundGray:(id)sender;
+- (IBAction)setBackgroundLight:(id)sender;
+- (IBAction)toggleSkybox:(id)sender;
+- (IBAction)toggleIBL:(id)sender;
 @end