+++ test/main.cpp
@@ -80,7 +80,7 @@ static const char *kGltfWithMesh = R"({
 // ---------------------------------------------------------------------------
 
 TEST(VersionTest, ReturnsExpectedVersion) {
-    EXPECT_EQ(systems::leal::campello_renderer::getVersion(), "0.1.4");
+    EXPECT_EQ(systems::leal::campello_renderer::getVersion(), "0.2.0");
 }
 
 TEST(VersionTest, VersionIsNonEmpty) {